[Entity Analytics] Move v1 evals into evals/v1/ subdir
Per @ymao1's review feedback: instead of adding `testIgnore` to v1's
Playwright config, mirror the v2 layout by moving v1 specs into an
`evals/v1/` subdirectory. The v1 `playwright.config.ts` now points at
`evals/v1` directly, symmetric with `playwright.v2.config.ts` already
pointing at `evals/v2`.

This avoids adding `testIgnore` plumbing to `@kbn/evals`'s
`createPlaywrightEvalsConfig`, keeps the package's public API surface
unchanged, and makes the directory structure self-documenting so future
v1 specs land in the correct folder.

- Revert `create_playwright_eval_config.ts` (no `testIgnore` option).
- Move 9 v1 specs from `evals/` to `evals/v1/` (git rename preserves
  blame).
- Update relative imports in moved specs: `../src/` -> `../../src/`.
- Update `playwright.config.ts` `testDir` to `evals/v1`.

Refs elastic/security-team#17138
