+++ x-pack/solutions/security/packages/kbn-evals-suite-entity-analytics/evals/v1/risk_score_engine_off.spec.ts
@@ -6,7 +6,7 @@
  */
 
 import { tags } from '@kbn/scout-security';
-import { evaluate } from '../src/evaluate';
+import { evaluate } from '../../src/evaluate';
 
 evaluate.describe(
   'SIEM Entity Analytics Skill - Risk Score Tests Without Data',
