+++ x-pack/solutions/security/packages/kbn-evals-suite-entity-analytics/playwright.config.ts
@@ -7,12 +7,7 @@
 import { createPlaywrightEvalsConfig } from '@kbn/evals';
 
 export default createPlaywrightEvalsConfig({
-  testDir: `${__dirname}/evals`,
-  // The v2 suite (`playwright.v2.config.ts`) owns specs under `evals/v2/` and
-  // requires the `evals_entity_analytics_v2` Scout configSet. Excluding them
-  // here prevents the v1 run from picking them up via recursive testDir walk
-  // and failing them under the wrong configSet.
-  testIgnore: ['**/v2/**'],
+  testDir: `${__dirname}/evals/v1`,
   repetitions: 1,
   timeout: 60 * 60_000,
 });
